requirements.txt
selenium==4.25.0
chromedriver-binary==93.0.4577.63
requests
gspread==4.0.1
pandas==1.3.3
beautifulsoup4==4.10.0
google-oauth==1.0.1
Flask==2.0.1
gunicorn==20.1.0
webdriver-manager >= 4.0.2
python-dotenv
Werkzeug==2.0.0
# zyte-smartproxy-selenium==1.0.10
blinker==1.7.0
selenium-wire
selenium_stealth
undetected-chromedriver